requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
requests==2.26.0
oauthlib==3.1.1
python-dotenv==0.19.0
PyJWT==2.3.0
python-dateutil==2.8.2
argparse==1.4.0